requirements.txt
importlib_metadata==6.0.0
locustio==0.999
numpy==1.24.2
pandas==1.5.3
Pillow==9.5.0
Pillow==9.5.0
prophet==1.1.2
streamlit==1.21.0
streamlit_extras==0.2.7
matplotlib
plotly
cython
